Adding a Pause Lock to the config file
If the Pause Lock is set to true in the config, the CS will not automatically provision boxes.
This setting overrides the pause button.
 So if the pauselock is set to true, nothing
will provision even if the pause button is used.
This was done so users could use the CS
to manually create/destroy boxes or use the api to create/destroy boxes.
It will also stop a user from accidently starting the CS to automatically provision boxes.

As the CS is started, it will check the config to see if the pauseLock is set to 'true'
If it is, a WARN will show 'Pause Lock engaged. Automatic Capacity will not occur.'
If pauseLock is set to 'false,' the user will see 'Automatic Capacity will occur unless paused in the UI.'
The pauseLock setting can not be changed by the API, it has to be manually set in the config before the CS
is started. This was done so that users can not accidently turn the setting off.
If the user pauses the CS via the UI, 'Service is paused in the UI.' is now displayed in the log.
If they user has both the pauseLock and pause set to true in the config, nothing is written to the log so
not to fill up disk space over time.